Notify user about the cleanup
Since experiments/fuzzing/fuzz.sh is running two syz-manager
instances in parallel, it uses temporary resources and
we delete those resources at the end of the script.
But since there is no information about it by the script at
the end, it seems like the script hangs, but in reality
it cleans up the temporary resources in the background.