Fix: Fix program hang due to tee and wait waiting each other
Since no arguments were provided to 'wait', it would wait for all
children of the script's process to exit.
However, one of these child processes was 'tee' which logs the script,
which also waits for the script to exit to concude its work.
Therefore, wait and tee would wait for each other and make the program
hang.
To avoid this, I got pids of each instances of fuzzing_experiments.sh
and provided them to wait.